Refactor change indicator work into a helper trait
